Work with the method
Now the img element with club crest gets an extra properties -
data-club-id(it is club ID from the API). It will help me with handling crest's click. I mean that when the user clicks on one of the crest, value of this property will be passed to another class which will send request to the appropriate resource, and (from this resource) gets the informations about the team.