Add a new property and correct the typo
This property is the representation of the place where are all
crests. I will use it to crest click handling -
https://kursjs.pl/kurs/events/events.php#dynamic-elements - The crests
are adding dynamically so I can't add listener directly to them. What is
more I correct a small typo(I could not split this change so I have
decided to put this two changes in one commit)